refactor(components) : reorganize home folder
