+++ src/components/quiz/DetailQuiz.tsx
@@ -1,5 +1,5 @@
 'use client'
-import MCQView from '@/components/home/components/MCQ_view';
+import MCQView from '@/components/quiz/mcq/MCQ_view';
 import { MCQ } from '../../model/interface/MCQ';
 
 import { useRouter } from 'next/navigation';
